Removed delivery team leads from about page and updated headshots for consulting leads
