+++ app/about/AboutPage.css
@@ -53,6 +53,7 @@
   width: 100%;
   aspect-ratio: 1 / 1;
   object-fit: cover;
+  object-position: top center;
   overflow: hidden;
 }
